get the profile location from the global gitconfig
I didn't want to pollute the global gitconfig with private user info in case
people want to include the file in their dotfiles. I also hated using a root
level config instead of putting it under .config. This seemed like a good way
of allowing the user to choose where to store the profile data.